renamed routes to plurals. Added routes for friendships
