+++ klaplongbois_api/index.js
@@ -7,13 +7,13 @@ server.get("/", (req, res, next) => {
 });
 
 // Load the user routes
-server.use('/api', require('./routes/v1/routes_user_v1'));
+server.use('/api', require('./routes/v1/routes_users_v1'));
 // Load the friendship routes
-server.use('/api', require('./routes/v1/routes_friendship_v1'));
+server.use('/api', require('./routes/v1/routes_friendships_v1'));
 // Load the thread routes
-server.use('/api', require('./routes/v1/routes_thread_v1'));
+server.use('/api', require('./routes/v1/routes_threads_v1'));
 // Load the comment routes
-server.use('/api', require('./routes/v1/routes_comment_v1'));
+server.use('/api', require('./routes/v1/routes_comments_v1'));
 
 server.listen(port, () => {
     console.log("Server is listening on port " + port);
